requirements/base.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
uvicorn==0.22.0
pytest==7.3.2
SQLAlchemy==2.0.16
pluggy==1.0.0
pydantic==1.10.9
pytest-asyncio==0.21.0
python-dotenv==1.0.0
python-multipart==0.0.6
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
orjson==3.9.1
packaging==23.1
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
ujson==5.8.0
watchfiles==0.19.0
websockets==11.0.3